% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brexcelfy.R
\name{brexcelfy}
\alias{brexcelfy}
\title{Emulate PHE Data Handling}
\usage{
brexcelfy(version = "Excel2003", nrow = NULL, ncol = NULL, verbose = TRUE)
}
\arguments{
\item{version}{Different truncation presets}

\item{nrow}{User defined number of rows to keep}

\item{ncol}{User defined number of columns to keep}

\item{verbose}{Logical, switch on/off function feedback}
}
\description{
This function truncates all data frames and matrices in your global environment to
emulate Public Health England (PHE) data handling with Excel (i.e. truncate them)
}
\details{
Available presets are \code{"Excel2003"} (default) and \code{"Excel2007"}.
The preset \code{"Excel2003"} keeps 65,536 rows and 255 columns, whereas
\code{"Excel2007"} keeps 1,048,576 rows and 16,384 columns.
Any other truncation is applied with the \code{nrow} and \code{ncol} options.

WARNING: RUN THIS FUNCTION ONLY, IF YOU REALLY WANT TO TRUNCATE ALL OBJECTS!
REMEMBER, THERE IS NO WAY TO RESTORE THE TRUNCATED PARTS, OTHER THAN TO
RECREATE THE TRUNCATED OBJECTS. DO NOT BLAME ME FOR DATA LOSS!
}
\examples{
ds1 <- matrix(1,ncol=100, nro=100)
ds2 <- as.data.frame(ds1)
ds3 <- matrix(1, ncol=3, nrow=10)
ds4 <- matrix(1, ncol=10, nrow=2)
ds5 <- matrix(1,ncol=3, nrow=2)
ds6 <- matrix(1,ncol=5, nrow=10)
brexcelfy(ncol=5, nrow=5)

}
